+++ tests/browser-console/src/log.ts
@@ -1,0 +1,5 @@
+export function invokeConsoleLog() {
+  console.log('inner:console.log');
+  console.time('inner:duration');
+  console.timeEnd('inner:duration');
+}