+++ packages/coverage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xbell/coverage
 
+## 0.8.6
+
+### Patch Changes
+
+- ed17157: Displays the browser print to the node console
+- Updated dependencies [ed17157]
+  - @xbell/color@0.8.6
+
 ## 0.8.5
 
 ### Patch Changes